+++ packages/vseed/src/pipeline/advanced/pipeline/funnel.ts
@@ -1,0 +1,28 @@
+import type { AdvancedPipeline } from 'src/types'
+import {
+  autoDimensions,
+  autoMeasures,
+  initAdvancedVSeed,
+  vchartBaseConfig,
+  vchartTheme,
+  pivotAdapter,
+  reshapeTo1D1M,
+  pivotReshapeTo1D1M,
+  annotation,
+  locale,
+  encodingFunnel,
+  funnelConfig,
+} from '../pipes'
+
+export const funnelAdvancedPipeline: AdvancedPipeline = [
+  initAdvancedVSeed,
+  locale,
+  autoMeasures,
+  autoDimensions,
+  pivotAdapter([reshapeTo1D1M], [pivotReshapeTo1D1M]),
+  encodingFunnel,
+  funnelConfig,
+  vchartBaseConfig,
+  vchartTheme,
+  annotation,
+]