+++ packages/vseed/src/pipeline/utils/measures/index.ts
@@ -1,1 +1,1 @@
-export { findMeasureById } from './findMeasureById'
+export * from './findMeasureById'