+++ packages/vseed/src/types/chartType/scatter/scatter.ts
@@ -1,0 +1,208 @@
+import type { Locale } from '../../i18n'
+import { zLocale } from '../../i18n'
+import {
+  zAnnotationArea,
+  zAnnotationHorizontalLine,
+  zAnnotationPoint,
+  zAnnotationVerticalLine,
+  zBackgroundColor,
+  zColor,
+  zCrosshairLine,
+  zDataset,
+  zDimensions,
+  zLabel,
+  zLegend,
+  zMeasures,
+  zPointStyle,
+  zTheme,
+  zTooltip,
+  zXLinearAxis,
+  zYLinearAxis,
+} from '../../properties'
+import type {
+  AnnotationArea,
+  AnnotationHorizontalLine,
+  AnnotationPoint,
+  AnnotationVerticalLine,
+  BackgroundColor,
+  Color,
+  Dataset,
+  Dimensions,
+  Label,
+  Legend,
+  Measures,
+  PointStyle,
+  Theme,
+  Tooltip,
+  CrosshairLine,
+  YLinearAxis,
+  XLinearAxis,
+} from '../../properties'
+import { z } from 'zod'
+
+/**
+ * 散点图类型定义
+ * @description 散点图，适用于展示数据的分布情况，通过点的位置表示数据的数值
+ * 适用场景:
+ * - 分析数据的分布特征, 如数据的中心趋势, 分布范围, 异常值等
+ * 数据要求:
+ * - 至少2个数值字段（度量）
+ * - 第一个指标字段会放至X轴, 其余指标会进行合并, 映射至Y轴
+ * - 指标名称和维度名称会合并, 作为图例项展示
+ * 默认开启的功能:
+ * - 默认开启图例、坐标轴、数据点标记、提示信息、趋势线
+ */
+export interface Scatter {
+  /**
+   * 散点图
+   * @description 散点图，适用于展示数据的分布情况，通过点的位置表示数据的数值
+   * @type {'scatter'}
+   * @example 'scatter'
+   */
+  chartType: 'scatter'
+  /**
+   * 数据集
+   * @description 符合TidyData规范的且已经聚合的数据集，用于定义图表的数据来源和结构, 用户输入的数据集并不需要进行任何处理, VSeed带有强大的数据重塑功能, 会自行进行数据重塑, 折线图的数据最终会被转换为2个维度, 1个指标.
+   * @type {Array<Record<string|number, any>>}
+   * @example [{month:'1月', value:100}, {month:'2月', value:150}, {month:'3月', value:120}]
+   */
+  dataset: Dataset
+
+  /**
+   * 维度
+   * @description 散点图的第一个维度被映射到X轴, 其余维度会与指标名称(存在多个指标时)合并, 作为图例项展示
+   * @type {Dimensions}
+   * @example [{id: "month", alias: "月份"}]
+   */
+  dimensions?: Dimensions
+
+  /**
+   * 指标
+   * @description 散点图的第一个指标字段会放至X轴, 其余指标会进行合并, 映射至Y轴
+   * @type {Measures}
+   * @example [{id: "value", alias: "数值"}]
+   */
+  measures?: Measures
+
+  /**
+   * 图表的背景颜色
+   * @default transparent 默认为透明背景
+   * @description 背景颜色可以是颜色字符串, 例如'red', 'blue', 也可以是hex, rgb或rgba'#ff0000', 'rgba(255,0,0,0.5)'
+   */
+  backgroundColor?: BackgroundColor
+
+  /**
+   * 颜色
+   * @description 颜色配置, 用于定义图表的颜色方案, 包括颜色列表, 颜色映射, 颜色渐变等.
+   */
+  color?: Color
+
+  /**
+   * 标签
+   * @description 标签配置, 用于定义图表的数据标签, 包括数据标签的位置, 格式, 样式等.
+   */
+  label?: Label
+
+  /**
+   * 图例
+   * @description 图例配置, 用于定义图表的图例, 包括图例的位置, 格式, 样式等.
+   */
+  legend?: Legend
+
+  /**
+   * 提示信息
+   * @description 提示信息配置, 用于定义图表的提示信息, 包括提示信息的位置, 格式, 样式等.
+   */
+  tooltip?: Tooltip
+
+  /**
+   * x轴
+   * @description 数值轴, x轴配置, 用于定义图表的x轴, 包括x轴的位置, 格式, 样式等.
+   */
+  xAxis?: XLinearAxis
+
+  /**
+   * y轴
+   * @description 数值轴, y轴配置, 用于定义图表的y轴, 包括y轴的位置, 格式, 样式等.
+   */
+  yAxis?: YLinearAxis
+
+  /**
+   * 垂直提示线
+   * @description  鼠标移动到图表上时, 显示的垂直提示线
+   */
+  crosshairLine?: CrosshairLine
+
+  /**
+   * 图表的主题, 主题是优先级较低的功能配置, 包含所有图表类型共用的通用配置, 与单类图表类型共用的图表配置
+   * @default light 默认为亮色主题
+   * @description 内置light与dark两种主题, 用户可以通过Builder自定义主题
+   * @example 'dark'
+   * @example 'light'
+   * @example 'customThemeName'
+   */
+  theme?: Theme
+
+  /**
+   * 点图元样式
+   * @description 点图元样式配置, 用于定义图表的点图元样式, 包括点图元的颜色, 边框等.
+   * 支持全局样式或条件样式配置
+   * 数据筛选器
+   * 若配置selector, 提供数值 selector, 局部数据 selector, 条件维度 selector, 条件指标 selector 共四类数据匹配能力
+   * 若未配置selector, 则样式全局生效.
+   */
+  pointStyle?: PointStyle | PointStyle[]
+
+  /**
+   * 标注点
+   * @description 标注点配置, 根据选择的数据, 定义图表的标注点, 包括标注点的位置, 格式, 样式等.
+   */
+  annotationPoint?: AnnotationPoint | AnnotationPoint[]
+
+  /**
+   * 标注垂直线
+   * @description 标注垂直线配置, 根据选择的数据, 定义图表的标注垂直线, 包括标注垂直线的位置, 样式等.
+   */
+  annotationVerticalLine?: AnnotationVerticalLine | AnnotationVerticalLine[]
+
+  /**
+   * 标注水平线
+   * @description 标注水平线配置, 根据选择的数据, 定义图表的标注水平线, 包括标注水平线的位置, 样式等.
+   */
+  annotationHorizontalLine?: AnnotationHorizontalLine | AnnotationHorizontalLine[]
+
+  /**
+   * 标注区域
+   * @description 标注区域配置, 根据选择的数据, 定义图表的标注区域, 包括标注区域的位置, 样式等.
+   */
+  annotationArea?: AnnotationArea | AnnotationArea[]
+
+  /**
+   * 语言
+   * @description 图表语言配置, 支持'zh-CN'与'en-US'两种语言, 另外可以调用 intl.setLocale('zh-CN') 方法设置语言
+   * @default 'zh-CN'
+   */
+  locale?: Locale
+}
+
+export const zScatter = z.object({
+  chartType: z.literal('scatter'),
+  dataset: zDataset.optional(),
+  dimensions: zDimensions.optional(),
+  measures: zMeasures.optional(),
+  backgroundColor: zBackgroundColor.optional(),
+  color: zColor.optional(),
+  label: zLabel.optional(),
+  legend: zLegend.optional(),
+  tooltip: zTooltip.optional(),
+  xAxis: zXLinearAxis.optional(),
+  yAxis: zYLinearAxis.optional(),
+  crosshairLine: zCrosshairLine.optional(),
+  theme: zTheme.optional(),
+  pointStyle: zPointStyle.optional(),
+  annotationPoint: zAnnotationPoint.optional(),
+  annotationVerticalLine: zAnnotationVerticalLine.optional(),
+  annotationHorizontalLine: zAnnotationHorizontalLine.optional(),
+  annotationArea: zAnnotationArea.optional(),
+  locale: zLocale.optional(),
+})